+++ mysql_rep_change.py
@@ -543,7 +543,6 @@ def main():
 
     """
 
-    cmdline = gen_libs.get_inst(sys)
     dir_chk_list = ["-d"]
     func_dict = {"-M": move_slave, "-R": move_slave, "-S": move_slave_up}
     opt_con_req_list = {"-M": ["-m", "-n"], "-R": ["-m", "-n"],
@@ -558,7 +557,7 @@ def main():
                "ssl_verify_id": "bool", "ssl_verify_cert": "bool"}
 
     # Process argument list from command line.
-    args_array = arg_parser.arg_parse2(cmdline.argv, opt_val_list)
+    args_array = arg_parser.arg_parse2(sys.argv, opt_val_list)
 
     if not gen_libs.help_func(args_array, __version__, help_message) \
        and not arg_parser.arg_require(args_array, opt_req_list) \
@@ -567,7 +566,7 @@ def main():
        and not arg_parser.arg_dir_chk_crt(args_array, dir_chk_list):
 
         try:
-            proglock = gen_class.ProgramLock(cmdline.argv,
+            proglock = gen_class.ProgramLock(sys.argv,
                                              args_array.get("-y", ""))
             run_program(args_array, func_dict, slv_key=slv_key)
             del proglock